[ANDROSDK-521] Remove UserParentCall and CategoryParent call
